pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "being"
version = "0.1.0"
description = "Open-source real-time talking head generation engine powered by 3D Gaussian Splatting"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9,<3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "numpy>=1.24.0,<2.0.0",
    "opencv-python>=4.8.0",
    "pillow>=9.5.0",
    "scipy>=1.10.0",
    "soundfile>=0.12.0",
    "librosa>=0.10.0",
    "imageio>=2.31.0",
    "imageio-ffmpeg>=0.4.8",
    "python-multipart>=0.0.6",
    "aiofiles>=23.0.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[project.scripts]
being = "being.cli:main"

[tool.setuptools.packages.find]
include = ["being*"]

[tool.ruff]
target-version = "py39"
line-length = 120
